requirements/build.txt
onnx==1.17.0
mmengine
wandb==0.20.1
numba==0.61.2
numpy==1.26.0
scipy==1.15.3
onnxconverter-common==1.14.0
pytorch-metric-learning==2.8.1
torch-geometric==2.6.1
xtcocotools==1.14.3
motmetrics==1.4.0
tabulate==0.9.0
supervision==0.25.1
seaborn==0.13.2
opencv-contrib-python==4.6.0.66
turbojpeg==0.0.2
